+++ packages/orq-rc/RELEASES.md
@@ -3149,3 +3149,13 @@ Based on:
 - [typescript v3.12.0-rc.2] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.2] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.2 - packages/orq-rc
+
+## 2025-08-20 11:39:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.0 (2.686.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.4 - packages/orq-rc